Task15. Finished realization of all article generation strategies. Modified process of generation
