+++ src/ArticleGeneration/Strategy/PlusGenerationStrategy.php
@@ -2,19 +2,10 @@
 
 namespace App\ArticleGeneration\Strategy;
 
-use App\Entity\Article;
-
 /**
  * Генерация статьи для пользователя с подпиской PLUS
  */
 class PlusGenerationStrategy extends BaseStrategy
 {
 
-    /**
-     * @inheritDoc
-     */
-    public function generate(Article $article): string
-    {
-        // TODO: Implement generate() method.
-    }
 }
